We have completed data analysis using Python , jupyter notebook, pandas, numpy and visualization
McDonald's is hip all over the world, including in India. However, the McDonald's menu is not universal. In India, there are some menu items that are similar, but others that are much different from the U.S. menu. Through comparing the McDonald's menu in the U.S. and India, we can see that fast food menus can truly give us a perspective on cultural and economic similarities and differences throughout the world.